+++ NEWS.md
@@ -11,6 +11,7 @@
 * New default tidier for `ggcoef_model()`, now using
   `broom.helpers::tidy_with_broom_or_parameters()` (#432)
 * Re-export methods from and redirect vignettes to the `{ggstats}` package (#452, #457)
+* Replaced ..scaled.. with after_stat(scaled) in ggscatmat (#467)
 
 
 # GGally 2.1.2